+++ mne/simulation/evoked.py
@@ -5,17 +5,19 @@
 # License: BSD (3-clause)
 import copy
 import warnings
+import math
 
 import numpy as np
 
 from ..io.pick import pick_channels_cov
 from ..forward import apply_forward
-from ..utils import check_random_state, verbose, _time_mask
+from ..utils import check_random_state, verbose, _time_mask, warn, deprecated
 
 
 @verbose
-def simulate_evoked(fwd, stc, info, cov, snr=3., tmin=None, tmax=None,
-                    iir_filter=None, random_state=None, verbose=None):
+def simulate_evoked(fwd, stc, info, cov, nave=3., tmin=None, tmax=None,
+                    iir_filter=None, random_state=None, verbose=None,
+                    snr=None):
     """Generate noisy evoked data.
 
     .. note:: No projections from ``info`` will be present in the
@@ -34,15 +36,16 @@ def simulate_evoked(fwd, stc, info, cov, snr=3., tmin=None, tmax=None,
         Measurement info to generate the evoked.
     cov : Covariance object
         The noise covariance.
-    snr : float
-        signal to noise ratio in dB. It corresponds to
-        10 * log10( var(signal) / var(noise) ).
+    nave : int
+        Number of averaged epochs.
+
+        .. versionadded:: 0.15.0
     tmin : float | None
         start of time interval to estimate SNR. If None first time point
-        is used.
+        is used. tmin is deprecated and will be removed in 0.16.
     tmax : float | None
         start of time interval to estimate SNR. If None last time point
-        is used.
+        is used. tmax is deprecated and will be removed in 0.16.
     iir_filter : None | array
         IIR filter coefficients (denominator) e.g. [1, -1, 0.2].
     random_state : None | int | np.random.RandomState
@@ -50,6 +53,11 @@ def simulate_evoked(fwd, stc, info, cov, snr=3., tmin=None, tmax=None,
     verbose : bool, str, int, or None
         If not None, override default verbose level (see :func:`mne.verbose`
         and :ref:`Logging documentation <tut_logging>` for more).
+    snr : float
+        signal to noise ratio in dB. It corresponds to
+        10 * log10( var(signal) / var(noise) ).
+
+        snr is deprecated and will be removed in 0.16.
 
     Returns
     -------
@@ -64,16 +72,46 @@ def simulate_evoked(fwd, stc, info, cov, snr=3., tmin=None, tmax=None,
 
     Notes
     -----
+    To make the equivalence between snr and nave, when the snr is given
+    instead of nave:
+        nave = (1 / 10 ** ((actual_snr - snr)) / 20) ** 2
+        where actual_snr is the snr to the generated noise before scaling.
+
     .. versionadded:: 0.10.0
     """
+    if snr is not None:
+        warn('snr is deprecated and will be removed in 0.16. Set it to '
+             'None to remove this warning. Set nave parameter instead.',
+             DeprecationWarning)
+
+    if tmin is not None:
+        warn('tmin is deprecated and will be removed in 0.16. Set it to '
+             'None to remove this warning.',
+             DeprecationWarning)
+
+    if tmax is not None:
+        warn('tmax is deprecated and will be removed in 0.16. Set it to '
+             'None to remove this warning.',
+             DeprecationWarning)
+
     evoked = apply_forward(fwd, stc, info)
-    if snr < np.inf:
+    if nave < np.inf:
         noise = simulate_noise_evoked(evoked, cov, iir_filter, random_state)
-        evoked_noise = add_noise_evoked(evoked, noise, snr, tmin=tmin,
-                                        tmax=tmax)
-    else:
-        evoked_noise = evoked
-    return evoked_noise
+
+        # Convert snr to nave before deprecation
+        if snr == np.inf:
+            nave = snr
+        elif snr is not None:
+            tmask = _time_mask(evoked.times, tmin, tmax,
+                               sfreq=evoked.info['sfreq'])
+            tmp = \
+                10 * np.log10(np.mean((evoked.data[:, tmask] ** 2)) /
+                              np.mean((noise.data ** 2)))
+            nave = 1. / 10 ** ((tmp - float(snr)) / 10)
+
+        evoked.data += noise.data / math.sqrt(nave)
+        evoked.nave = np.int(nave)
+    return evoked
 
 
 def simulate_noise_evoked(evoked, cov, iir_filter=None, random_state=None):
@@ -132,11 +170,11 @@ def _generate_noise(info, cov, iir_filter, random_state, n_samples, zi=None):
     return noise, zf
 
 
+@deprecated('add_noise_evoked will be deprecated and removed in 0.16.')
 def add_noise_evoked(evoked, noise, snr, tmin=None, tmax=None):
     """Add noise to evoked object with specified SNR.
 
     SNR is computed in the interval from tmin to tmax.
-
     Parameters
     ----------
     evoked : Evoked object
@@ -150,7 +188,6 @@ def add_noise_evoked(evoked, noise, snr, tmin=None, tmax=None):
         start time before event
     tmax : float
         end time after event
-
     Returns
     -------
     evoked_noise : Evoked object
@@ -158,8 +195,8 @@ def add_noise_evoked(evoked, noise, snr, tmin=None, tmax=None):
     """
     evoked = copy.deepcopy(evoked)
     tmask = _time_mask(evoked.times, tmin, tmax, sfreq=evoked.info['sfreq'])
-    tmp = 10 * np.log10(np.mean((evoked.data[:, tmask] ** 2).ravel()) /
-                        np.mean((noise.data ** 2).ravel()))
+    tmp = 10 * np.log10(np.mean((evoked.data[:, tmask] ** 2)) /
+                        np.mean((noise.data ** 2)))
     noise.data = 10 ** ((tmp - float(snr)) / 20) * noise.data
     evoked.data += noise.data
     return evoked